+++ composer.json
@@ -129,7 +129,9 @@
             "EonX\\EasyActivity\\Tests\\": "packages/EasyActivity/tests",
             "EonX\\EasyActivity\\Tests\\Bridge\\Symfony\\Fixtures\\App\\": "packages/EasyActivity/tests/Bridge/Symfony/Fixtures/app/src",
             "EonX\\EasyApiPlatform\\Tests\\": "packages/EasyApiPlatform/tests",
+            "EonX\\EasyApiPlatform\\Tests\\Application\\": "packages/EasyApiPlatform/tests/Application/src",
             "EonX\\EasyApiPlatform\\Tests\\Fixture\\App\\": "packages/EasyApiPlatform/tests/Fixture/app/src",
+            "EonX\\EasyApiPlatform\\Tests\\Unit\\": "packages/EasyApiPlatform/tests/Unit/src",
             "EonX\\EasyApiToken\\Tests\\": "packages/EasyApiToken/tests",
             "EonX\\EasyAsync\\Tests\\": "packages/EasyAsync/tests",
             "EonX\\EasyAsync\\Tests\\Unit\\": "packages/EasyAsync/tests/Unit/src",
